test: convert type tests to .test.ts so knip recognizes them
knip flagged the two *.test-d.ts files as unused — the vitest include
globs only tests/**/*.test.ts. Fold createModel type assertions into the
existing tests/extend-adapter.test.ts and rename the compatible type test
to compatible-types.test.ts.
