+++ packages/ai-openai/tests/compatible-types.test.ts
