+++ packages/ai/tests/extend-adapter.test.ts
@@ -10,6 +10,7 @@
 import { describe, expect, expectTypeOf, it } from 'vitest'
 import { createModel, extendAdapter } from '../src/extend-adapter'
 import { BaseTextAdapter } from '../src/activities/chat/adapter'
+import type { ExtendedModelDef } from '../src/extend-adapter'
 import { chat } from '../src/activities/chat'
 import { EventType } from '../src/types'
 import type { StreamChunk, TextOptions } from '../src/types'
@@ -250,5 +251,43 @@ describe('extendAdapter', () => {
       const adapter = extendedMock('mock-gpt-4')
       expect(adapter.model).toBe('mock-gpt-4')
     })
+  })
+})
+
+describe('createModel capabilities overload', () => {
+  it('(name, inputArray) still infers name + input (backward compat)', () => {
+    const m = createModel('legacy-model', ['text', 'image'])
+    expectTypeOf(m.name).toEqualTypeOf<'legacy-model'>()
+    expectTypeOf(m.input).toEqualTypeOf<readonly ['text', 'image']>()
+  })
+
+  it('(name, capabilities) captures features and tools', () => {
+    const m = createModel('reasoner', {
+      input: ['text'],
+      features: ['reasoning', 'structured_outputs'],
+      tools: ['web_search'],
+    })
+    expectTypeOf(m.name).toEqualTypeOf<'reasoner'>()
+    expectTypeOf(m.input).toEqualTypeOf<readonly ['text']>()
+    expectTypeOf(m.features).toEqualTypeOf<
+      readonly ['reasoning', 'structured_outputs'] | undefined
+    >()
+    expectTypeOf(m.tools).toEqualTypeOf<readonly ['web_search'] | undefined>()
+  })
+
+  it('capabilities-form model is still an ExtendedModelDef', () => {
+    const m = createModel('reasoner', { input: ['text'] })
+    expectTypeOf(m).toMatchTypeOf<ExtendedModelDef>()
+  })
+
+  it('maps providerOptions to modelOptions', () => {
+    interface MyOpts {
+      reasoningEffort: 'low' | 'high'
+    }
+    const m = createModel('reasoner', {
+      input: ['text'],
+      providerOptions: {} as MyOpts,
+    })
+    expectTypeOf(m.modelOptions).toEqualTypeOf<MyOpts>()
   })
 })
